requirements.txt

# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
requests  

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Environment Variables
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# LangChain & RAG
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.2

# LLM Integration 
langchain-groq==0.0.1
groq==0.4.1

# Vector Database
chromadb==0.4.22

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4

# Embeddings
sentence-transformers==2.3.1
openai==1.7.2

# Streamlit UI
streamlit==1.31.0
streamlit-extras==0.3.6
extra-streamlit-components==0.1.71
plotly==5.18.0
requests==2.31.0

# Utilities
tiktoken==0.5.2
rich==13.7.0

# Testing (Optional for now)
pytest==7.4.4
httpx==0.26.0